test encapsulation and compilation
