+++ src/main/java/NexusPublisher.java
@@ -1,11 +1,6 @@
-import java.io.IOException;
-import java.io.UncheckedIOException;
-import java.nio.file.FileVisitResult;
-import java.nio.file.FileVisitor;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.nio.file.attribute.PosixFilePermissions;
 import java.util.Arrays;
 import java.util.Map;
@@ -38,7 +33,7 @@ class NexusPublisher {
             e.printStackTrace();
             System.exit(1);
         } finally {
-            deleteDir(gnupghomeDir);
+            DirectoryDeleter.deleteDir(gnupghomeDir);
         }
     }
     
@@ -61,44 +56,5 @@ class NexusPublisher {
             throw new IllegalArgumentException("Needs environment variable '" + envName + "' to be defined and non-blank. See readme!");
         }
         return value;
-    }
-
-    /**
-     * Deletes directory recursively.
-     *
-     * @param dir the directory to delete
-     */
-    private static void deleteDir(Path dir) {
-        try {
-            if (dir == null) {
-                return;
-            }
-
-            Files.walkFileTree(dir, new FileVisitor<Path>() {
-                @Override
-                public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-                  return FileVisitResult.CONTINUE;
-                }
-    
-                @Override
-                public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                  Files.delete(file);
-                  return FileVisitResult.CONTINUE;
-                }
-    
-                @Override
-                public FileVisitResult visitFileFailed(Path file, IOException exc) throws IOException {
-                  return FileVisitResult.CONTINUE;
-                }
-    
-                @Override
-                public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-                  Files.delete(dir);
-                  return FileVisitResult.CONTINUE;
-                }
-              });
-        } catch (IOException e) {
-            throw new UncheckedIOException("Failed to delete directory " + dir, e);
-        }
     }
 }
